# lib/fs: Path handling optimizations