Two independent optimizations to path handling. Benchmark results on Linux/amd64, Go 1.17.1:

    # lib/osutil
    name                old time/op    new time/op    delta
    TraversesSymlink-8    8.86µs ± 3%    8.53µs ± 4%  -3.79%  (p=0.000 n=18+20)
    
    name                old alloc/op   new alloc/op   delta
    TraversesSymlink-8    1.06kB ± 0%    1.06kB ± 0%    ~     (all equal)
    
    name                old allocs/op  new allocs/op  delta
    TraversesSymlink-8      15.0 ± 0%      15.0 ± 0%    ~     (all equal)

    # lib/fs, new benchmark
    name                          old time/op    new time/op    delta
    WindowsInvalidFilenameValid-8     875ns ± 1%     150ns ± 1%  -82.84%  (p=0.000 n=9+9)
    WindowsInvalidFilenameNUL-8       276ns ± 4%     121ns ± 3%  -56.26%  (p=0.000 n=10+10)
    
    name                          old alloc/op   new alloc/op   delta
    WindowsInvalidFilenameValid-8     32.0B ± 0%     16.0B ± 0%  -50.00%  (p=0.000 n=10+10)
    WindowsInvalidFilenameNUL-8       32.0B ± 0%     19.0B ± 0%  -40.62%  (p=0.000 n=10+10)
    
    name                          old allocs/op  new allocs/op  delta
    WindowsInvalidFilenameValid-8      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=10+10)
    WindowsInvalidFilenameNUL-8        2.00 ± 0%      2.00 ± 0%     ~     (all equal)